+++ package.json
@@ -13,8 +13,6 @@
     "ejs": "^3.1.9",
     "express": "^4.18.2",
     "express-session": "^1.17.3",
-    "multer": "^1.4.5-lts.1",
-    "nodemon": "^2.0.22",
     "socket.io": "^4.6.2",
     "sqlite3": "^5.1.6"
   }
